refactor: move getEnv() back from @internal/utils
