+++ .changeset/config.json
@@ -2,7 +2,9 @@
   "$schema": "https://unpkg.com/@changesets/config@2.3.1/schema.json",
   "changelog": "@changesets/cli/changelog",
   "commit": ["@changesets/cli/commit", { "skipCI": false }],
-  "fixed": [["@avalabs/evm-module", "@avalabs/avalanche-module", "@avalabs/vm-module-types", "@avalabs/bitcoin-module"]],
+  "fixed": [
+    ["@avalabs/evm-module", "@avalabs/avalanche-module", "@avalabs/vm-module-types", "@avalabs/bitcoin-module"]
+  ],
   "linked": [],
   "access": "restricted",
   "baseBranch": "main",
